implement savedconfig & add arm64 qemu

+++ sys-boot/u-boot/u-boot-2020.07_rc4.ebuild
@@ -17,8 +17,6 @@ SLOT="0"
 KEYWORDS="-* ~arm64"
 IUSE="build-tfa"
 
-TFADIR="/usr/share/trusted-firmware-a"
-
 DEPEND="
 	sys-apps/dtc
 	!build-tfa? ( sys-firmware/trusted-firmware-a-bin:= )
@@ -26,43 +24,67 @@ DEPEND="
 RDEPEND="${DEPEND}"
 BDEPEND=""
 S="${WORKDIR}/${MY_P}"
+TFADIR="/usr/share/trusted-firmware-a"
 
 src_configure() {
-	emake ${U_BOOT_CONFIG} || die
+	if use savedconfig; then
+		restore_config .config
+	else
+		emake ${U_BOOT_CONFIG} || die
+	fi
 }
 
 src_compile() {
 	#Safe bet for now
 	unset CFLAGS CXXFLAGS CPPFLAGS LDFLAGS
 
+	#Uboot seems very gcc-centric, clang support will be... adventurous
 	export CROSS_COMPILE="${CHOST}-"
 
-	case ${U_BOOT_CONFIG} in
-		rock64-rk3328_defconfig)
-			emake BL31=${TFADIR}/rk3328/bl31.elf || die
-			;;
-		rockpro64-rk3399_defconfig|pinebook-pro-rk3399_defconfig)
-			emake BL31=${TFADIR}/rk3399/bl31.elf || die
-			;;
-		pine64-lts_defconfig|pine64_plus_defconfig|pinebook_defconfig|pine_h64_defconfig)
-			emake BL31=${TFADIR}/sun50i_a64/bl31.elf || die
-			;;
-		rpi_4_defconfig)
-			emake BL31=${TFADIR}/rpi4/bl31.elf || die
-			;;
-		**)
-			eerror "unsupported platform"
-			eerror "please submit a bug if you want your platform supported"
-			;;
-	esac
+	if use savedconfig; then
+		emake BL31="${BL31PATH}" || die
+		einfo "built u-boot with savedconfig"
+	else
+		case ${U_BOOT_CONFIG} in
+			qemu_arm64_defconfig)
+				emake || die
+				;;
+			rock64-rk3328_defconfig)
+				emake BL31=${TFADIR}/rk3328/bl31.elf || die
+				;;
+			rockpro64-rk3399_defconfig|pinebook-pro-rk3399_defconfig)
+				emake BL31=${TFADIR}/rk3399/bl31.elf || die
+				;;
+			pine64-lts_defconfig|pine64_plus_defconfig|pinebook_defconfig|pine_h64_defconfig)
+				emake BL31=${TFADIR}/sun50i_a64/bl31.elf || die
+				;;
+			rpi_4_defconfig)
+				emake BL31=${TFADIR}/rpi4/bl31.elf || die
+				;;
+			**)
+				eerror "unsupported platform"
+				eerror "please submit a bug if you want your platform supported"
+				;;
+		esac
+		einfo "built u-boot with ${U_BOOT_CONFIG} configuration"
+	fi
 
-	einfo "built u-boot with ${U_BOOT_CONFIG} configuration"
 }
 
 src_install() {
 	insinto /usr/share/u-boot
 
 	case ${U_BOOT_CONFIG} in
+		qemu_arm64_defconfig)
+			# no idea which one of those are necessary, let's just give em all for now
+			doins u-boot
+			doins u-boot.bin
+			doins u-boot.lds
+			doins u-boot.map
+			doins u-boot-nodtb.bin
+			doins u-boot.srec
+			doins u-boot.sym
+			;;
 		rock64-rk3328_defconfig)
 			;;
 		rockpro64-rk3399_defconfig|pinebook-pro-rk3399_defconfig)
@@ -81,19 +103,24 @@ src_install() {
 }
 
 pkg_postinst() {
-	elog "To install U-Boot:"
+	if ! use savedconfig; then
+		elog "To install U-Boot:"
 
-	case ${U_BOOT_CONFIG} in
-		rock64-rk3328_defconfig)
-			;;
-		rockpro64-rk3399_defconfig|pinebook-pro-rk3399_defconfig)
-			elog "dd if=/usr/share/u-boot/idbloader.img of=/dev/... seek=64 conv=notrunc"
-			elog "dd if=/usr/share/u-boot/u-boot.itb of=/dev/... seek=16384 conv=notrunc"
-			elog "where /dev/... can either be the SD or eMMC"
-			;;
-		pine64-lts_defconfig|pine64_plus_defconfig|pinebook_defconfig|pine_h64_defconfig)
-			;;
-		rpi_4_defconfig)
-			;;
-	esac
+		case ${U_BOOT_CONFIG} in
+			qemu_arm64_defconfig)
+				# Documentation on this on the wiki would be helpful, TBD
+				;;
+			rock64-rk3328_defconfig)
+				;;
+			rockpro64-rk3399_defconfig|pinebook-pro-rk3399_defconfig)
+				elog "dd if=/usr/share/u-boot/idbloader.img of=/dev/... seek=64 conv=notrunc"
+				elog "dd if=/usr/share/u-boot/u-boot.itb of=/dev/... seek=16384 conv=notrunc"
+				elog "where /dev/... can either be the SD or eMMC"
+				;;
+			pine64-lts_defconfig|pine64_plus_defconfig|pinebook_defconfig|pine_h64_defconfig)
+				;;
+			rpi_4_defconfig)
+				;;
+		esac
+	fi
 }
